refactor/Factor out commo code at function getUserAndSessionContextFuture
